Analysis of viscosity text files
This script loads all viscosity data files in a folder and calculates the moving-window standard deviation of the data, and segments the raw data into "steady" and "unsteady" portions by comparing the moving window standard deviation at each point to the standard deviation threshold specified by the user. The steady and unsteady viscosities are plotted along with the standard deviation of viscosity, and the mean, standard deviation, and bounds of the 95% confidence interval of viscosity are saved within a .mat file.

This code can only be used if the spindle speed was kept constant throughout the trial. The spindle type output only applies if 60 RPM was
used with the 40 spindle and 120 RPM was used with the 51 spindle.

This code also only works with filenames similar to '25C_07-11-16_hum_la###_trial###' OR '10C_02-10-17_dry_water_trial###'